Move test deck ballot generation utils to libs/utils
