+++ apps/admin/frontend/package.json
@@ -63,7 +63,6 @@
     "@votingworks/types": "workspace:*",
     "@votingworks/ui": "workspace:*",
     "@votingworks/utils": "workspace:*",
-    "array-unique": "^0.3.2",
     "assert": "^2.0.0",
     "base64-js": "1.5.1",
     "browserify-zlib": "^0.2.0",
@@ -109,7 +108,6 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/array-unique": "^0.3.0",
     "@types/base64-js": "^1.3.0",
     "@types/connect": "^3.4.35",
     "@types/debug": "4.1.8",
